requirements.txt
streamlit>=1.18
openai>=1.0.0
python-dotenv
google-generativeai
